+++ pkg/key/appcatalogentry.go
@@ -7,3 +7,7 @@ import (
 func AppCatalogEntryManagedBy(projectName string) string {
 	return fmt.Sprintf("%s-unique", projectName)
 }
+
+func AppCatalogEntryName(catalogName, appName, appVersion string) string {
+	return fmt.Sprintf("%s-%s-%s", catalogName, appName, appVersion)
+}