+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   "config.giantswarm.io/major-version" annotation.
 - Set "config-controller.giantswarm.io/version" label to "0.0.0" on created App
   CRs.
+- Validate App CR configmap and secret names if managed by config-controller.
 
 ## [3.5.0] - 2020-11-27