+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Validate App CR configmap and secret names if managed by config-controller.
 
+### Changed
+
+- Change `config.giantswarm.io/major-version` annotation to `config.giantswarm.io/version`.
+
 ## [3.6.0] - 2020-12-01
 
 ### Added